refactor(models): move models to utils/models directory
- Move all model files from strands_tools/models/ to strands_tools/utils/models/
- Update import statements in graph.py, think.py, use_agent.py, and workflow.py
- Reorganize codebase structure for better organization
