+++ src/strands_tools/utils/models/anthropic.py
