+++ src/strands_tools/utils/models/bedrock.py
