+++ src/strands_tools/utils/models/litellm.py
