+++ src/strands_tools/utils/models/model.py
