+++ src/strands_tools/utils/models/openai.py
