+++ src/strands_tools/utils/models/writer.py
